[refactor]  앱 구조 분리 및 오류 수정
