+++ moodico/users/forms.py
@@ -1,7 +1,6 @@
 from django import forms
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib.auth.models import User
-from .models import *
 
 class CustomSignupForm(UserCreationForm):
     username = forms.CharField(
@@ -35,8 +34,3 @@ class CustomSignupForm(UserCreationForm):
         help_texts = {
             'username': '',
         }
-
-class UploadForm(forms.ModelForm):
-    class Meta:
-        model = Upload
-        fields = ['image_path']
